LV: Changes in hooks, state management, and PoC of simpler tree
