+++ src/navigation/NavigationSidebar.tsx
@@ -1,32 +1,14 @@
-import React, { useEffect, useRef, useState } from "react";
-import { useHistory } from "react-router-dom";
 import { History } from "history";
-import { HashLink as Link } from "react-router-hash-link";
-import { createContainer } from "unstated-next";
-
-import * as Styles from "./NavStyles";
-import init from "./nav";
-
-import "./mmenu-overrides.css";
-import ApiSidebar from "./ApiSidebar";
+import React, { useEffect, useState } from "react";
+import { useHistory } from "react-router-dom";
 import useBrowserEffect from "src/util/useBrowserEffect";
-// import "mmenu-js/dist/mmenu.css"
-
-import styled from "styled-components";
-import { DropdownChild, MenuItemProps } from "./components/DropdownChild";
-import { DropdownParent, MenuParentProps } from "./components/DropdownParent";
-import {
-  SubcategoryList,
-  SubcategoryProps,
-} from "./components/SubcategoryList";
-import { CoreCategory } from "./components/CoreCategory";
-import { IconSVGDiv } from "./components/styled";
-import {
-  bullHornIcon,
-  dropdownCaret,
-  graphGrowthIcon,
-  graphGrowthIconBig,
-} from "./components/icons";
+import { createContainer } from "unstated-next";
+import { CoreCategoryController } from "./components/CoreCategory";
+import { bullHornIcon, graphGrowthIconBig } from "./components/icons";
+import "./mmenu-overrides.css";
+import init, { MMenuID } from "./nav";
+import * as Styles from "./NavStyles";
+import { learningSaasquatchSubCategories } from "./portalNavItems";
 
 /* Growth Automation menu items */
 
@@ -34,36 +16,18 @@ import {
 
 /* MMenu Stuff */
 function useMMenu() {
-  const [mmenuApi, setMMenuApi] = useState(null);
-  return {
-    get mmenuApi() {
-      return mmenuApi;
-    },
-    set mmenuApi(next) {
-      setMMenuApi(next);
-    },
-  };
-}
-
-export const MMenuContext = createContainer(useMMenu);
-
-export const modalRoot =
-  typeof document === "undefined" ? undefined : document.createElement("div");
-
-/* Rendering Function */
-export function NavigationSidebar() {
   const history: History<any> = useHistory();
-  // const container = useRef(null);
-  const mmenu = MMenuContext.useContainer();
 
-  useBrowserEffect(() => {
-    mmenu.mmenuApi = init(modalRoot, history);
-  }, [modalRoot]);
+  const [mmenuApi, setMMenuApi] = useState(null);
+  const [activePages, setActivePages] = useState<string[]>([]);
 
   /* Clicked list item changes, will be applied to StyledLink */
-  const [currentPage, setcurrentPage] = useState<string>("/");
-  const [activePages, setActivePages] = useState<string[]>([]);
+  const path = history.location.pathname;
+  const hash = history.location.hash;
+  const pathAndHash = path + hash;
+  const currentPage = pathAndHash;
   const isActive = (pageKey: string) => activePages.includes(pageKey);
+  const clearActivePages = () => setActivePages([]);
 
   const toggleActivePage = (pageKey: string) => {
     setActivePages((prev) =>
@@ -72,9 +36,6 @@ export function NavigationSidebar() {
         : addPageKey(prev, pageKey)
     );
   };
-  function clearActivePages() {
-    setActivePages([]);
-  }
 
   const removePageKey = (keys: string[], pageKey: string) =>
     keys.filter((key) => key !== pageKey);
@@ -97,177 +58,56 @@ export function NavigationSidebar() {
     };
   }, []);
 
-  useEffect(() => {
-    const path = history.location.pathname;
-    const hash = history.location.hash;
-    const pathAndHash = path + hash;
-    setcurrentPage(pathAndHash);
-  }, [history.location.pathname, history.location.hash]);
-
-  const adminPortalItems: MenuItemProps[] = [
-    {
-      path: "/success/using-referral-saasquatch/",
-      title: "Using the SaaSquatch Portal",
-      currentPage,
+  return {
+    get mmenuApi() {
+      return mmenuApi;
     },
-    {
-      path: "/success/navigating-the-portal/",
-      title: "Navigating the SaaSquatch Portal",
-      currentPage,
+    set mmenuApi(next) {
+      setMMenuApi(next);
     },
-    {
-      path: "/success/referral-feed/",
-      title: "The Referral Feed",
-      currentPage,
-    },
-    {
-      path: "/features/analytics/",
-      title: "Program Analytics",
-      currentPage,
-    },
-  ];
-
-  const growthAutoItems: MenuItemProps[] = [
-    {
-      path: "/growth/ga-101",
-      title: "Growth Automation 101",
-      currentPage,
-    },
-    {
-      path: "/growth/customer-lifecycle",
-      title: "Growth Automation Customer Lifecycle",
-      currentPage,
-    },
-    {
-      path: "/growth/saasquatch-ga",
-      title: "SaaSquatch Growth Automation Platform",
-      currentPage,
-    },
-  ];
-
-  const referralProgramsItems: MenuItemProps[] = [
-    {
-      path: "/success/intro/",
-      title: "Referral Programs 101",
-      currentPage,
-    },
-    {
-      path: "/success/referral-program-optimization/",
-      title: "Referral Program Optimization",
-      currentPage,
-    },
-    {
-      path: "/success/core-topics/",
-      title: "The SaaSquatch Referral Program Loop",
-      currentPage,
-    },
-    {
-      path: "/success/touchpoints/",
-      title: "Referral Marketing Channels",
-      currentPage,
-    },
-    {
-      path: "/success/referral-program-retargeting/",
-      title: "Referral Program Retargeting",
-      currentPage,
-    },
-    {
-      path: "/success/share-options/",
-      title: "Referral Program Sharing Options",
-      currentPage,
-    },
-  ];
-
-  const fraudSecurityManageItems: MenuItemProps[] = [
-    {
-      path: "/success/referral-security/",
-      title: "Security Management System",
-      currentPage,
-    },
-    {
-      path: "/fraud-and-security/",
-      title: "Fraud, Security & Fake Referrals",
-      currentPage,
-    },
-  ];
-
-  const learningSaasquatchDropdowns: MenuParentProps[] = [
-    {
-      title: "SaaSquatch Admin Portal",
-      parentID: "adminPortal",
-      menuItems: adminPortalItems,
-      svgIcon: dropdownCaret,
-      toggleActivePage,
-      isActive,
-    },
-    {
-      title: "Growth Automation",
-      parentID: "growthAuto",
-      menuItems: growthAutoItems,
-      svgIcon: dropdownCaret,
-      toggleActivePage,
-      isActive,
-    },
-    {
-      title: "Referral Programs",
-      parentID: "referralPrograms",
-      menuItems: referralProgramsItems,
-      svgIcon: dropdownCaret,
-      toggleActivePage,
-      isActive,
-    },
-    {
-      title: "Fraud and Security Management",
-      parentID: "fraudSecurityManage",
-      menuItems: fraudSecurityManageItems,
-      svgIcon: dropdownCaret,
-      toggleActivePage,
-      isActive,
-    },
-  ];
-
-  const learningSaasquatchSubCategories: SubcategoryProps = {
-    title: "Learning SaaSquatch",
-    path: "/success/",
     currentPage,
-    svgIcon: graphGrowthIcon,
-    dropdowns: learningSaasquatchDropdowns,
+    activePages,
+    isActive,
+    toggleActivePage,
+    clearActivePages,
   };
+}
+
+export const MMenuContext = createContainer(useMMenu);
+
+export const modalRoot =
+  typeof document === "undefined" ? undefined : document.createElement("div");
+
+/* Rendering Function */
+export function NavigationSidebar() {
+  const history: History<any> = useHistory();
+  // const container = useRef(null);
+  const mmenu = MMenuContext.useContainer();
+
+  useBrowserEffect(() => {
+    mmenu.mmenuApi = init(modalRoot, history);
+  }, [modalRoot]);
 
   return (
     <Styles.Container>
-      <nav id="my-menu">
+      <nav id={MMenuID}>
         <ul className="baseMenu">
-          <CoreCategory
+          <CoreCategoryController
             title="Product Updates"
             path="/product-news"
-            currentPage={currentPage}
             icon={bullHornIcon}
             hasNextPage={false}
-            setActivePages={setActivePages}
           />
 
-          <CoreCategory
+          <CoreCategoryController
             title="Learning SaaSquatch"
             path="/success/"
-            currentPage={currentPage}
             icon={graphGrowthIconBig}
             hasNextPage={true}
             subcategoryList={learningSaasquatchSubCategories}
-            setActivePages={setActivePages}
           />
         </ul>
       </nav>
     </Styles.Container>
   );
-}
-
-export interface SVGProps {
-  fill?: string;
-  width?: string;
-  height?: string;
-  viewBox?: string;
-  d?: string;
-  clicked?: boolean;
-  dropdownSelected?: boolean;
 }
