+++ src/navigation/components/styled.tsx
@@ -1,6 +1,7 @@
 import React from "react";
-import styled from "styled-components";
 import { HashLink as Link } from "react-router-hash-link";
+import styled from "styled-components";
+import { SVGProps } from "../types/SVGProps";
 
 export const StyledLink = styled(Link)`
   display: flex !important;
@@ -204,7 +205,7 @@ export const IconSVGDiv = styled.div`
   width: 25px;
   height: auto;
 `;
-export const SidebarSVGIcon: React.FC<SVGProps> = ({
+export const SidebarSVGIcon: React.FC<SVGProps & SVGState> = ({
   width,
   height = "auto",
   viewBox,
@@ -227,12 +228,7 @@ export const SidebarSVGIcon: React.FC<SVGProps> = ({
     </IconSVGDiv>
   );
 };
-export interface SVGProps {
-  fill?: string;
-  width?: string;
-  height?: string;
-  viewBox?: string;
-  d?: string;
+export interface SVGState {
   clicked?: boolean;
   dropdownSelected?: boolean;
 }
