+++ src/navigation/types/MenuItemProps.tsx
@@ -1,0 +1,5 @@
+
+export interface MenuItemProps {
+  path: string;
+  title: string;
+}
