+++ src/navigation/types/MenuParentProps.tsx
@@ -1,0 +1,9 @@
+import { MenuItemProps } from "./MenuItemProps";
+import { SVGProps } from "./SVGProps";
+
+export interface MenuParentProps {
+  title: string;
+  parentID: string;
+  menuItems: MenuItemProps[];
+  svgIcon: SVGProps;
+}
