+++ src/navigation/types/SVGProps.tsx
@@ -1,0 +1,10 @@
+/**
+ * Defines what an SVG looks like
+ */
+export interface SVGProps {
+  fill?: string;
+  width?: string;
+  height?: string;
+  viewBox?: string;
+  d?: string;
+}
