+++ src/navigation/types/SubcategoryProps.tsx
@@ -1,0 +1,9 @@
+import { MenuParentProps } from "./MenuParentProps";
+import { SVGProps } from "./SVGProps";
+
+export interface SubcategoryProps {
+  title: string;
+  path: string;
+  svgIcon: SVGProps;
+  dropdowns: MenuParentProps[];
+}
